+++ src/main/java/kitchenpos/menu/domain/dto/response/MenuProductResponse.java
@@ -1,7 +1,5 @@
 package kitchenpos.menu.domain.dto.response;
 
-import kitchenpos.menu.domain.MenuProduct;
-
 public class MenuProductResponse {
 
     private Long seq;
