+++ src/main/java/kitchenpos/ordertable/domain/dto/request/OrderTableChangeEmptyRequest.java
